+++ flies-rest-api/src/main/java/org/fedorahosted/flies/rest/dto/DataHook.java
@@ -15,7 +15,7 @@ import javax.xml.bind.annotation.XmlType;
 public class DataHook extends AbstractBaseResource implements Resource {
 
 	private String id;
-	private long version = 1;
+	private Integer version = 1;
 	
 	private List<Object> extensions;
 	
@@ -36,11 +36,11 @@ public class DataHook extends AbstractBaseResource implements Resource {
 	}
 
 	@XmlAttribute(name="version", required=true)
-	public long getVersion() {
+	public Integer getVersion() {
 		return version;
 	}
 	
-	public void setVersion(long version) {
+	public void setVersion(Integer version) {
 		this.version = version;
 	}
