+++ flies-rest-api/src/main/java/org/fedorahosted/flies/rest/dto/Reference.java
@@ -15,7 +15,7 @@ import javax.xml.bind.annotation.XmlType;
 public class Reference extends AbstractBaseResource implements Resource{
 	
 	private String id;
-	private long version = 1;
+	private Integer version = 1;
 	
 	private List<Object> extensions;
 	
@@ -41,11 +41,11 @@ public class Reference extends AbstractBaseResource implements Resource{
 	}
 
 	@XmlAttribute(name="version", required=true)
-	public long getVersion() {
+	public Integer getVersion() {
 		return version;
 	}
 	
-	public void setVersion(long version) {
+	public void setVersion(Integer version) {
 		this.version = version;
 	}
